+++ Jarvis.Framework.Tests/ProjectionsTests/InsertEvent.cs
@@ -1,5 +1,7 @@
+using Fasterflect;
 using Jarvis.Framework.Shared.Events;
 using Jarvis.Framework.TestHelpers;
+using System;
 
 namespace Jarvis.Framework.Tests.ProjectionsTests
 {
@@ -9,6 +11,14 @@ namespace Jarvis.Framework.Tests.ProjectionsTests
         public InsertEvent()
         {
             this.AssignIdForTest(new MyAggregateId(1));
+        }
+
+        public InsertEvent(Int64 checkpointToken, Int64 aggregateVersion, Int32 eventPosition)
+        {
+            this.AssignIdForTest(new MyAggregateId(1));
+            this.SetPropertyValue("CheckpointToken", checkpointToken);
+            this.SetPropertyValue("EventPosition", eventPosition);
+            this.SetPropertyValue("Version", aggregateVersion);
         }
     }
 }