+++ Wiki/ReleaseNotes.md
@@ -2,6 +2,13 @@
 
 ## vNext
 
+## 7.1.0
+
+- Optimization for pollers (allows user to use Watch)
+- Some optimization for concurrent checkpoint tracker.
+
+## 7.0.0
+
 - Added time to live for TrackedMessageModel.
 
 # 6.6.2